reorganize files into clients/models/routes
