+++ gantry/clients/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+from .gitlab import GitlabClient
+from .prometheus import PrometheusClient
